Move System/Platform logic to single location to allow bypass
